refactor(iam): providers.py 1154줄을 패키지로 분해
기존 한 파일에 AWS / K8s / LDAP / GCP / Azure 5종 + 공통 + dispatcher가
다 있어 책임 경계가 흐릿했음. 패키지로 쪼개 provider별 격리.

iam/providers/
  __init__.py   — public API re-export (기존 import path 그대로)
  base.py       — FetchedIdentity / FetchedPermission / FetchResult /
                  AttachResult dataclass + resolve_credentials + logger
  aws.py        — boto3 기반. fetch / attach / detach / _stub
  k8s.py        — kubernetes SDK. RBAC API + Core API. RoleBinding 부여
  ldap.py       — ldap3. AD / OpenLDAP / FreeIPA. 그룹 멤버십으로 부여
  gcp.py        — google-cloud-resource-manager + iam_admin. 프로젝트 IAM
  azure.py      — azure-mgmt-authorization. roleAssignments
  registry.py   — CAPABILITIES 매트릭스 + fetch_for / attach_for / detach_for
                  dispatcher. if/elif 체인 대신 dict lookup

호출처 변경 0건. `from app.iam.providers import fetch_for, ...`
그대로 동작 (__init__.py에서 re-export).

검증:
- public API import OK
- get_capabilities() → 6 entries
- fetch_for(AWS stub) → 4 identities / 4 permissions, stub=True
- /api/v1/iam/capabilities 200 OK
